+++ mycli/completion_refresher.py
@@ -1,6 +1,8 @@
 import threading
 from typing import Callable
 
+import pymysql
+
 from mycli.packages.special.main import COMMANDS
 from mycli.packages.sqlresult import SQLResult
 from mycli.sqlcompleter import SQLCompleter
@@ -58,22 +60,25 @@ class CompletionRefresher:
 
         # Create a new sqlexecute method to populate the completions.
         e = sqlexecute
-        executor = SQLExecute(
-            e.dbname,
-            e.user,
-            e.password,
-            e.host,
-            e.port,
-            e.socket,
-            e.character_set,
-            e.local_infile,
-            e.ssl,
-            e.ssh_user,
-            e.ssh_host,
-            e.ssh_port,
-            e.ssh_password,
-            e.ssh_key_filename,
-        )
+        try:
+            executor = SQLExecute(
+                e.dbname,
+                e.user,
+                e.password,
+                e.host,
+                e.port,
+                e.socket,
+                e.character_set,
+                e.local_infile,
+                e.ssl,
+                e.ssh_user,
+                e.ssh_host,
+                e.ssh_port,
+                e.ssh_password,
+                e.ssh_key_filename,
+            )
+        except pymysql.err.OperationalError:
+            return
 
         # If callbacks is a single function then push it into a list.
         if callable(callbacks):